Split configuration serialization compile unit
